requirements.txt
django==4.2.19
djangorestframework==3.14.0
google-generativeai==0.3.2
pinecone-client==2.2.4
python-dotenv==1.0.0
redis==5.0.1